refactor(agent): extract composer controls
